refactor(core): own changed-file refresh logic
